pyproject.toml
[project]
name = "agent-orchestrator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.92.0",
    "langgraph>=0.3.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
